Add a bazel rule for python library generation and reorganize test code
The goal is to be able to have multiple integration tests comparing the
behaviour between `go` and `Python` (or other languages). The main
concept which we introduce here is the separation of the protocol
definition and the library and the test itself. Ideally, per one
scenario we should have a single test target. This will allow bazel to
rerun things when they do change.

Summary:
- Add a new macro for generating a py_zserio_library.
- Add a new macro for generating test payload together with a generator
  CLI, which requires a `data.py` file to contain a constructor for the
  data to be output.
- Remove godog and simplify how tests are defined.
- Move the zserio library generation next to the schema.
