+++ BUILD.bazel
@@ -5,6 +5,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 # gazelle:prefix github.com/woven-planet/go-zserio
 # gazelle:exclude example/**
 # gazelle:exclude **/*_gen.zs.go
+#
+# For generated zserio libraries:
+# gazelle:resolve go gen/github.com/woven-planet/go-zserio/test/go/reference_modules/testobject1/testobject //testdata/reference_modules:go_testobject
 gazelle(name = "gazelle")
 
 gazelle(
