--- test/go/BUILD.bazel
@@ -1,38 +1,0 @@
-load("//:rules.bzl", "go_zserio_library")
-
-rootpkg = "gen/github.com/woven-planet/go-zserio/test/go"
-
-go_zserio_library(
-    name = "core_types",
-    srcs = [
-        "//testdata:zserio_schema_all",
-    ],
-    pkg = "reference_modules.core.types",
-    rootpackage = rootpkg,
-)
-
-go_zserio_library(
-    name = "core_instantiations",
-    srcs = [
-        "//testdata:zserio_schema_all",
-    ],
-    pkg = "reference_modules.core.instantiations",
-    rootpackage = rootpkg,
-    deps = [
-        ":core_types",
-    ],
-)
-
-go_zserio_library(
-    name = "testobject",
-    srcs = [
-        "//testdata:zserio_schema_all",
-    ],
-    pkg = "reference_modules.testobject1.testobject",
-    rootpackage = rootpkg,
-    visibility = ["//visibility:public"],
-    deps = [
-        ":core_instantiations",
-        ":core_types",
-    ],
-)
