+++ testdata/BUILD.bazel
@@ -1,30 +1,9 @@
 load("@bazel_skylib//rules:build_test.bzl", "build_test")
 load("//:rules.bzl", "go_zserio_library")
 
-exports_files(
-    srcs = glob(["*.zs"]),
-    visibility = ["//internal:__subpackages__"],
-)
-
 filegroup(
     name = "examples",
     srcs = glob(["*.zs"]),
-    visibility = ["//visibility:public"],
-)
-
-filegroup(
-    name = "zserio_schema",
-    srcs = [
-        "reference_modules/all.zs",
-    ],
-    visibility = ["//visibility:public"],
-)
-
-filegroup(
-    name = "zserio_schema_all",
-    srcs = glob([
-        "reference_modules/*/*.zs",
-    ]),
     visibility = ["//visibility:public"],
 )
